Move autojudge submodule
