+++ judges/__init__.py
@@ -1,7 +1,6 @@
-from judges.autojudge.core import AutoJudge
 from judges.base import Jury
 from judges.classifiers import *
 from judges.graders import *
 
-__all__ = ["Jury", "AutoJudge"]
+__all__ = ["Jury"]
