+++ judges/classifiers/auto/__init__.py
@@ -1,0 +1,3 @@
+from judges.classifiers.auto.core import AutoJudge
+
+__all__ = ["AutoJudge"]
